+++ test/UriTest.php
@@ -1,24 +1,23 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-validator for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-validator/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-validator/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Validator;
+namespace LaminasTest\Validator;
 
-use Zend\Validator;
-use Zend\Uri\Exception\InvalidArgumentException;
+use Laminas\Uri\Exception\InvalidArgumentException;
+use Laminas\Validator;
 
 /**
- * @group      Zend_Validator
+ * @group      Laminas_Validator
  */
 class UriTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * @var \Zend\Validator\Uri
+     * @var \Laminas\Validator\Uri
      */
     protected $validator;
 
@@ -36,14 +35,14 @@ class UriTest extends \PHPUnit_Framework_TestCase
     {
         $validator = $this->validator;
         $uriHandler = $validator->getUriHandler();
-        $this->assertInstanceOf('Zend\Uri\Uri', $uriHandler);
+        $this->assertInstanceOf('Laminas\Uri\Uri', $uriHandler);
         $this->assertTrue($validator->getAllowRelative());
         $this->assertTrue($validator->getAllowAbsolute());
     }
 
     public function testConstructorWithArraySetsOptions()
     {
-        $uriMock = $this->getMock('Zend\Uri\Uri');
+        $uriMock = $this->getMock('Laminas\Uri\Uri');
         $validator = new Validator\Uri(array(
             'uriHandler' => $uriMock,
             'allowRelative' => false,
@@ -56,7 +55,7 @@ class UriTest extends \PHPUnit_Framework_TestCase
 
     public function testConstructorWithArgsSetsOptions()
     {
-        $uriMock = $this->getMock('Zend\Uri\Uri');
+        $uriMock = $this->getMock('Laminas\Uri\Uri');
         $validator = new Validator\Uri($uriMock, false, false);
         $this->assertEquals($uriMock, $validator->getUriHandler());
         $this->assertFalse($validator->getAllowRelative());
@@ -86,7 +85,7 @@ class UriTest extends \PHPUnit_Framework_TestCase
         $allowAbsolute, $allowRelative, $isAbsolute, $isRelative, $isValid, $expects
     ) {
         $uriMock = $this->getMock(
-            'Zend\Uri\Uri',
+            'Laminas\Uri\Uri',
             array('parse', 'isValid', 'isAbsolute', 'isValidRelative')
         );
         $uriMock->expects($this->once())
@@ -105,7 +104,7 @@ class UriTest extends \PHPUnit_Framework_TestCase
 
     public function testUriHandlerThrowsExceptionInParseMethodNotValid()
     {
-        $uriMock = $this->getMock('Zend\Uri\Uri');
+        $uriMock = $this->getMock('Laminas\Uri\Uri');
         $uriMock->expects($this->once())
             ->method('parse')
             ->will($this->throwException(new InvalidArgumentException()));
@@ -133,12 +132,12 @@ class UriTest extends \PHPUnit_Framework_TestCase
 
     public function testUriHandlerCanBeSpecifiedAsString()
     {
-        $this->validator->setUriHandler('Zend\Uri\Http');
-        $this->assertInstanceOf('Zend\Uri\Http', $this->validator->getUriHandler());
+        $this->validator->setUriHandler('Laminas\Uri\Http');
+        $this->assertInstanceOf('Laminas\Uri\Http', $this->validator->getUriHandler());
     }
 
     /**
-     * @expectedException Zend\Validator\Exception\InvalidArgumentException
+     * @expectedException Laminas\Validator\Exception\InvalidArgumentException
      */
     public function testUriHandlerStringInvalidClassThrowsException()
     {
@@ -146,7 +145,7 @@ class UriTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Zend\Validator\Exception\InvalidArgumentException
+     * @expectedException Laminas\Validator\Exception\InvalidArgumentException
      */
     public function testUriHandlerInvalidTypeThrowsException()
     {